service로 function 이동 및 소스 정리
