+++ .idea/misc.xml
@@ -16,7 +16,7 @@
         <entry key="..\:/random-chatting/app/src/main/res/layout/list_activity.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/slide_sample.xml" value="0.1" />
         <entry key="..\:/random-chatting/app/src/main/res/layout/test.xml" value="0.1" />
-        <entry key="..\:/random-chatting/app/src/main/res/layout/user_list_activity.xml" value="0.12" />
+        <entry key="..\:/random-chatting/app/src/main/res/layout/user_list_activity.xml" value="0.1" />
       </map>
     </option>
   </component>
