+++ app/src/main/res/layout/user_list_activity.xml
@@ -114,16 +114,17 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/user_list_btn_layout"
         android:layout_width="match_parent"
-        android:layout_height="100dp"
+        android:layout_height="wrap_content"
         android:gravity="left"
         app:layout_constraintTop_toBottomOf="@+id/maskImageLayout">
     <Button
         android:id="@+id/user_list_activity_btn_next_user"
         android:layout_width="200dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="100dp"
-        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
         android:text="다  음"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
